requirements.txt
python-dotenv==0.19.0
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
paho-mqtt==1.5.1
pycryptodome==3.10.1
python-dotenv==0.19.0
requests==2.26.0
tuya-iot-py-sdk==0.4.1
urllib3==1.26.6
certifi==2021.5.30
charset-normalizer==2.0.6
idna==3.2
paho-mqtt==1.5.1
pycryptodome==3.10.1
python-dotenv==0.19.0
requests==2.26.0
tuya-iot-py-sdk==0.4.1
urllib3==1.26.6
